requirements.txt
aiogram==3.10.0
apscheduler==3.10.4
peewee==3.17.5
python-dotenv==1.0.1